move cg solver to linearze
